+++ tests/presenters/test_login_in_company_presenter.py
@@ -7,8 +7,8 @@ from arbeitszeit_web.presenters.log_in_company_presenter import LogInCompanyPres
 from tests.session import FakeSession
 from tests.translator import FakeTranslator
 
+from ..forms import LoginForm
 from .dependency_injection import get_dependency_injector
-from .forms import LoginForm
 from .url_index import CompanyUrlIndexImpl
